Added USPS / UPS shipping tracking to SARAH
Created automations for the pond to turn the pond on/off 10 minutes
of every hour, unless manually activated in which case it turns on for
60 minutes.

Added an automation to turn on the garden pump to water the plants outside
for 20 minutes a day

Added a brand new two-switch relay to control the porch lighting / bar power
and automations to turn everything off at 2am automatically

Created a new sensor for the power monitoring in the house to automatically
estimate power usage in terms of $ /hr so I can feel even more guilty about
the amount of power we consume.

Added the backyard camera to SARAH and programmed it to store a snapshot
of what it sees every minute to S3 for Machine Learning Purposes.

Created a simple web interface to let me classify images from a given
S3 bucket into positive/negative to later be fed into a ML algorithm.