Adding a shell script to power a RTL433 to MQTT bridge (which allows Honeywell transmitters to be detected).
Implementing the honeywell transmitters, plus automations to notify us
if the doors are left open for too long.

Adding web links to various aspects of the house on other machines for convienence sake

Adding the Automatic device tracker from the car and the start of
some automations to map events coming from that service into MQTT
topics.